+++ lib/crypt.dart
@@ -1,5 +1,6 @@
-// Copyright (c) 2015, 2021 Hoylen Sue. All rights reserved. Use of this source code
-// is governed by a BSD-style license that can be found in the LICENSE file.
+// Copyright (c) 2015-2022 Hoylen Sue.
+// All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
 
 /// One-way string hashing for salted passwords using the Unix crypt format.